Added periodic task
