+++ main/main/settings.py
@@ -44,6 +44,7 @@ INSTALLED_APPS = [
     'allauth.account',
     'allauth.socialaccount',
     'allauth.socialaccount.providers.google',
+    # 'django_celery_beat',
 ]
 
 MIDDLEWARE = [
